[imager] Refactor internal names to fix golint
